pyproject.toml
[project]
name = "qc-parallelizer"
version = "1.3.0"
description = "A module for optimally combining and distributing quantum circuits"
authors = [
    { name = "Henri Ahola", email = "firstname.lastname@vtt.fi" }
]
readme = "README.pypi.md"
requires-python = ">= 3.10"
dependencies = [ "qiskit >= 1.0, < 1.3", "z3-solver >= 4.0.0.0", "rustworkx >= 0.16.0" ]

[project.urls]
GitHub = "https://github.com/vttresearch/qc-parallelizer"
Notebooks = "https://github.com/vttresearch/qc-parallelizer/tree/main/notebooks"

[project.optional-dependencies]
tests = [ "iqm-client[qiskit]", "pytest", "pytest-cov" ]
notebooks = [ "iqm-client[qiskit]", "ipykernel", "matplotlib", "pylatexenc" ]

[build-system]
requires = [ "setuptools>=68.0.0", "setuptools_scm[toml]>=7" ]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# See configuration details in https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = [ "F", "E", "W", "I001", "I002" ]
ignore = [ "E101", "W191", "F401", "F811", "E712" ]
# F401: ignore unused imports, breaks Pytest fixtures
# F811: ignore redefinitions, also breaks Pytest fixtures
# E712: ignore comparisons with False

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
# Pytest causes various DeprecationWarnings in Qiskit, so we ignore them
# Warnings that original from a call in our code is treated normally, though
# NOTE: is ignoring a good idea?
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.pyright]
exclude = ["**/notebooks", "**/.ipynb*"]
ignore = ["**/notebooks", "**/.ipynb*"]
